♻️ refactor: Redis 유틸리티 클래스를 RedisService로 리팩토링 및 테스트 코드 수정
- RedisUtil 클래스와 관련된 메서드들을 RedisService로 이전 및 리팩토링
- 이메일 인증 토큰 저장 및 조회 기능 추가
- RedisService에 대한 단위 테스트 추가 (set, get, delete, hasKey 등)
